Mudança do tipo da variavel PESO

+++ PastaFontes/Produto.java.java
@@ -3,23 +3,23 @@
 
 public class Produto {
 
-	private double peso;
+	private int peso;
 	private double altura;
 	
 	public Produto() {
 		
 	}
 	
-	public Produto(double peso, double altura) {
+	public Produto(int peso, double altura) {
 		this.peso=peso;
 		this.altura=altura;
 	}
 	
-	public double getPeso() {
+	public int getPeso() {
 		return peso;
 	}
 	
-	public void setPeso(double peso) {
+	public void setPeso(int peso) {
 		this.peso = peso;
 	}
